+++ e2e_test/lib/test_declOrder.dart
@@ -1,0 +1,7 @@
+class ChildClass extends ParentClass {
+
+}
+
+class ParentClass {
+
+}